# mcp/chatgpt-action-openapi.yaml
openapi: 3.1.0
info:
  title: SkyFlow Flight Booking Assistant API
  description: Live flight search, baggage policies, cancellation rules, and airport information API for ChatGPT Custom Actions and Plugins.
  version: 1.0.0
servers:
  - url: https://YOUR-NGROK-OR-PUBLIC-DOMAIN.ngrok-free.app
    description: Public HTTPS server for ChatGPT Cloud Actions
paths:
  /api/chatgpt/flights/search:
    get:
      operationId: searchFlights
      summary: Search available scheduled flights between origin and destination airports
      description: Fetches real-time flight schedules, airlines, ticket prices (INR), and seat availability between any two airport IATA codes.
      parameters:
        - name: origin
          in: query
          required: true
          schema:
            type: string
            example: DEL
          description: 3-letter IATA origin airport code (e.g. DEL, BOM, BLR, MAA, HYD, CCU, GOI)
        - name: destination
          in: query
          required: true
          schema:
            type: string
            example: BOM
          description: 3-letter IATA destination airport code (e.g. BOM, DEL, BLR, MAA, HYD, CCU, GOI)
      responses:
        '200':
          description: List of matching flights with prices and available seats
...